+++ commands_extra/ros_easy/shell/load_ros2.sh
@@ -17,6 +17,7 @@
 
 # echo $release
 # echo $version 
+# echo $ROS_DISTRO
 
 case $version in 
     "22.04")