+++ commands_extra/ros_easy/shell/load_tb3.sh
@@ -13,9 +13,24 @@
 # QQ Qun: 创客智造D群:562093920                               
 ################################################
 
-if [ -d ~/ros2_tb3_ws ]; then
-    source ~/ros2_tb3_ws/install/setup.bash
-    export ROS_DOMAIN_ID=30 #TURTLEBOT3
-    export GAZEBO_MODEL_PATH=$GAZEBO_MODEL_PATH:~/ros2_tb3_ws/src/turtlebot3/turtlebot3_simulations/turtlebot3_gazebo/models
-    export TURTLEBOT3_MODEL=burger
-fi
+case "$ROS_DISTRO" in
+    "noetic")
+        if [ -d ~/ros1_tb3_ws ]; then
+            source ~/ros1_tb3_ws/devel/setup.bash
+            export GAZEBO_MODEL_PATH=$GAZEBO_MODEL_PATH:~/ros1_tb3_ws/src/turtlebot3/turtlebot3_simulations/turtlebot3_gazebo/models
+
+        fi
+        export ROS_DOMAIN_ID=30 #TURTLEBOT3
+        export TURTLEBOT3_MODEL=burger
+        export LDS_MODEL=LDS-02 #LDS-01/LDS-02
+        ;;
+    *)
+        if [ -d ~/ros2_tb3_ws ]; then
+            source ~/ros2_tb3_ws/install/setup.bash
+            export GAZEBO_MODEL_PATH=$GAZEBO_MODEL_PATH:~/ros2_tb3_ws/src/turtlebot3/turtlebot3_simulations/turtlebot3_gazebo/models
+        fi
+        export ROS_DOMAIN_ID=30 #TURTLEBOT3
+        export TURTLEBOT3_MODEL=burger
+        export LDS_MODEL=LDS-02 #LDS-01/LDS-02
+        ;;
+esac